SaveManager Setup
Error: 'vec': is not a member of 'SaveObject::MySaveStruct'
This is not a bug, I'm just testing accessing member values- looking at making this a paramter tomorrow. This branch is in active development. This commit is being made to share progress between machines and create a backup in git history.